requirements.txt
absl-py==1.4.0
anyio==3.7.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
config==0.5.1
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
flatbuffers==23.3.3
gast==0.4.0
gensim==4.3.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
hangul-utils==0.2
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
jax==0.4.8
jedi==0.18.2
joblib==1.2.0
JPype1==1.4.1
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
konlpy==0.6.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.15.1
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seqeval==1.2.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
